accidently typed schema code in index.js instead of book.js, i moved over the code
